pyproject.toml
[tool.poetry]
name = "notify-python-demo"
version = "0.1.0"
description = ""
authors = ["Carly Jugler <carjug@gmail.com>"]
readme = "README.md"
packages = [{include = "notify_python_demo"}]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.21.0"
notifications-python-client = "^6.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
black = "^22.8.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
